backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "flask-caching>=2.3.1",
    "flask-cors>=6.0.2",
    "flask-jwt-extended>=4.7.1",
    "flask-mail>=0.10.0",
    "flask-migrate>=4.1.0",
    "flask-restful>=0.3.10",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "python-dotenv>=1.2.2",
]
